See: https://tools.suckless.org/dmenu/patches/center/
This patch centers dmenu in the middle of the screen.

Previously this has been achieved through the xyw patch and a bash
script to calculate the centered x and y positions. However, this is a
slow and overly complex way which is hard to integrate into programs
which call dmenu directly, eg surf or tabbed. On the other hand, This
small standalone patch is instantaneous and works globally.

With dmenu-center-20200111-8cd37e1.diff, you can use -c to center dmenu.